pyproject.toml
[project]
name = "siteup"
version = "0.1.0"
description = "Simple website monitoring tool"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "aiohttp>=3.9.0",
    "aiosqlite>=0.19.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings",
    "python-multipart>=0.0.6",
    "argon2-cffi",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend"] 